[rust png] Initial support for Bazel build.
After this CL, `bazelisk build experimental/rust_png` succeeds and
builds `SkPngRustCodec.cpp` and its transitive dependencies.

This CL updates the `WORKSPACE.bazel` dependency on `cxx` and
`cxxbridge-cmd` crates to version `1.0.128`.  The following commands
have been used to mirror the crates:

```
$ go run ./bazel/gcs_mirror/gcs_mirror.go --url https://crates.io/api/v1/crates/cxxbridge-cmd/1.0.128/download --sha256 1e3d98437a33d422ba30975d549f754d50300404096f17f5037abce2d91e3cfa --add_suffix=.tar.gz
$ go run ./bazel/gcs_mirror/gcs_mirror.go --url https://crates.io/api/v1/crates/cxx/1.0.128/download --sha256 54ccead7d199d584d139148b04b4a368d1ec7556a1d9ea2548febb1b9d49f9a4 --add_suffix=.tar.gz
```

Bug: chromium:356875275
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/909037
Auto-Submit: Łukasz Anforowicz <lukasza@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>

+++ experimental/rust_png/impl/SkPngRustCodec.h
@@ -70,7 +70,7 @@ private:
     Result incrementalDecode(DecodingState& decodingState, int* rowsDecoded);
 
     // Helper for seeking to the start of image data for the given frame.
-    Result seekToStartOfFrame(size_t index);
+    Result seekToStartOfFrame(int index);
 
     // Helper for reading until the start of the next `fdAT` sequence.
     Result readToStartOfNextFrame();
